+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add some design, and set the random theme at the center of screen
+
 ## [0.1.2] - 2022-10-27
 
 ### Fixed